Make it so that preload+fork does not add a stack layer
This is an insane use of a source filter...

The 'spawn' command is used to start the test runner, it preloads stuff,
then forks for each test, running the test. The problem is that .t files
expect the be the very first thing in the stack, that is caller(0) should
return nothing. preload + fork makes it pretty much impossible to avoid
adding a stack frame. This filter is able to avoid adding a stack frame
by preventing the rest of the 'yath' script from being read, and instead
it feeds lines from the test file into the parser.