Revamp watcher to do less event mangling
Previously the watcher and harness would mangle events in irriversable
ways before letting the renderer have them. It would also re-order some
events as necessary (buffered subtests). Ultimately this meant the final
event stream and the original event stream were transformed in ways that
meant the first was nto useful for inspection, and the later was not
useful for replay.

Now the watcher/harness is significantly more sane. Any changes the
watcher makes to events can be reversed. Events that need to be
re-ordered are found in both their original location and the new
location with flags that identify them (and prevent the wrong one from
being rendered).

The resulting log is useful both for inspection and replay. Things are
generally more sane as there is less state mutation, and the mutation
that is done can be reversed easily (as seen in Feeder/JSONL.pm)