Handle Formatter::Stream being initialized via !import
When a module runs `use Formatter::Stream` the PID gets locked into a
variable called ROOT_PID.  This is used to error out if/when the
formatter gets reinitialized in a child process.

However, in the case that T2_FORMATTER is set, nothing ever _imports_
Test2::Formatter::Stream, it is just created as the default formatter
and then made the formatter.  In this case, the best thing we can do
is just initialize ROOT_PID at this point.  Could we be erroneously in
a child process? Maybe.. or it could just be that something imported
Test::Builder early on in the loading process with T2_FORMATTER set.

This addresses issue #87:

    Test case:

    env T2_FORMATTER=Stream perl -MTest::Builder -E 'say "hi"';

    Expected result: all good
    Result: dies with an exception concerning ROOT_PID being undef