Standardize how fields are specified
We want to be able to pass custom fields onto display tools like yath-ui

For jobs this is done in events with the 'harness_job_fields' facet.
Require newer version of plugins to make sure we get it.

For runs this is done in plugins using the 'fields' attribute. Update
the git plugin to provide it.