Make sure -Dlib libs are loaded BEFORE ANYTHING
Yath accepts the -Dlib (--dev-lib=PATH) option. This option's guarentee
is that it will add the desired libs to the front of @INC before
ANYTHING is loaded. %INC should be blank until this is done.

Complications:
 * -D without an arg adds 'lib', 't/lib', 'blib/lib', 'blib/arch'
 * -Dpath or -D=path are how you specify something other than the
    defaults
 * -D options can come from .yath.rc and .yath.user.rc!
 * The .rc files have a rel(path) syntax to say path is relative to the
   .rc file
 * The main script MUST NOT contaminate the main:: namespace in any way
 * Due to how goto::file works for yath+preloading this all must be in a
   BEGIN.

All of this was previously handled by a util lib... the problem is that
util lib had to be loaded to use it, and for sanity the util lib loaded
things like File::Spec and Cmd. Problem is loading any of these kicks
off a chain of loads breaking the %INC is clean guarentee.

So, now:

 * The logic to find and parse RC files is part of the main script.
 * The -D options are parsed/processed in the main script.
 * The ugly relative paths thatmight break if we chdir are cleaned up
   using Cwd::realpath() AFTER being added to @INC, Cwd is loaded after
   the relative paths are in @INC, so any -D options can effect loading
   it or its deps.

I did not make a "clean" util class to load because the -D option to
yath is for yath/yath-plugin development and would need to load even
that clean lib from paths specified in -D, so chicken+egg....

Testing this of course is a must, but is also crazy. Testing that
nothing is loaded until it is done is also crazy as it meakes loading
test tools complicated. That said using comments and string evals
(shudder) I made it work. This test may be the wrong kind of 'clever',
but it works.