Fix race condition in rendering
The main loop in the test command that reads events and renders them is
non-blocking because we need to frequently call ipc->wait. This resulted
in a race condition where very rarely an incomplete event would be read.
This fixes the issue by buffering when an incomplete event is read.