Consistent signal setup
This is a regression noticed with the 'major_refactor'
and signal improvements.

As shown by the provided test 'signals/abrt_or_iot.t',
this is flapping when using yath, whereas it's consistent
when using prove.

The root of the issue is the 'local %SIG = %SIG' which
has multiple issues.

1/ 'local %SIG' does not work, signals are not restored
when exiting the scope.

2/ '%SIG = %SIG' is also bad as it forces a hash setup
in a random order due to hash keys ordering...

Several signals are using multiple names for the same id:
- ABRT and IOT
- CHLD and CLD
- IO and POLL
- SYS and UNUSED

When relying on the signal name has shown by 'abrt_or_iot.t'
the first signal set is the one giving its name to the signal,
even if the signal is undefined...

By forcing the signal names to be ordered in the same way they
are listed in Config, we provide a consistent behavior and will
always use the same signal name.