Do not use relative path in @INC
Using relative path in @INC could lead to unexpected behavior.
Code can be loaded from uncontrolled locations.

When setting @INC with devlib: 'lib', 'blib', 'blib/arch'
we should not use the relative PATH but prefer the absolute one.

Note that the workardound for 'projects' command need some extra love,
as this seems to be the expected behavior for projects, we should avoid
using an environment variable.