Overhaul the include processing system
 * Do not "clean" paths to be not relative until necessary, due to
   ch_dir functionality
 * Do not include -D paths when running tests, -D indicates paths for
   yath to look at for its own stuff, not for tests
 * Handle INC processing consistently via a util tool
 * lib/blib are added in that order after and -I's by default
 * lib/blib can be disabled (not new)
 * -l and -b can be used to manually specify one, disabling the other,
   or both can be specified. When manually specified the order is
   preserved along with -I's

Because we cannot make the paths absolute until late in the process, due
to ch_dir/projects command there is no simple/sane way to verify the
-I's and -l and -b paths exists when they are specified, so no warnings
will happen if the user specifies invalid paths. Prove/Perl also have no
such warnings so this is at least consistent.

I am not filtering invalid paths, by the time we are ready to do so it
is not possible to tell which paths were specified by the user, the
yathrc file, perl5lib, defaults, or others, and we do not want to filter
ones we did not add. So we WILL NOT be removing invalid dirs from @INC.

Ticket #147 should be closed after this, things have improved and been
as fixed as they will get.

If someone wants to make a PR with proper state management to track
which @INC entries came from where, and be capable of filtering only the
user-specified and/or default ones AFTER any chdir has been accounted
for, I will consider it.

I will also consider a patch to add a '--filter-inc' flag that does
remove any invalid dirs in the places where @INC is set (Runner.pm,
runner.pm, Job.pm) I do not personally feel it has any value.