Skip persist-based tests in automated testing
Any integration tests that launch yath in persistent mode have a high
chance of failure in automated testing based both on the fact that a
process is daemonized, and that write access to non-temp directories is
required. This commit makes the tests skip when AUTOMATED_TESTING is
set.

This also fixes the github workflow tests, so those are re-enabled now.