Fix __FILE__ value when running 'projects' yath command
The __FILE__ is set either by goto::file or by exec() when launching the
test.

In the goto::file case chdir() happened after goto::file() was called,
resulting in a relative path relative to the wrong cwd().

In the exec() case an absolute path was used, which is not terrible, but
is inconsistent.

In both cases you now get the relative path, relative to the current
directory.