Do not block on a process exit anymore
This block was originally put in place when the only reason it could be
reached was all job slots were full, at which point blocking on a
process exit (assuming there are any running processes) made sense.

Now it is poddible to reach this point due to a lack of necessary
resources. Resources can come in late or be blocked/available because of
external factors. It no longer makes sense to block on process exit
here.

Still doing the wait() for process cleanup reasons, but no longer
blocking.

Fixes #200