Overhaul Stream formatter and add ASCII to internal JSON
Stream has been overhauled to remove some hacks. Specifically it was
changed a while back to write files to a dir, starting a new file
whenever the process id or thread id changed. The only times the
formatter is used in a new process/thread is for buffered events in
subtest started in child processes.

Now the formatter only accepts buffered events in the main thread. This
does mean less verbosity when a buffered subtest in a child process
runs, but it makes handling things overall a lot simpler. It also means
we can have a single channel for events meaning pipes are not viable,
and files are not needed. As an extra bonus this will avoid some
permissions issues when tests change effective users.

Stream now takes an FD to send events to, if none is found it sends them
to STDOUT as fallback (same as before).

Any attemtp to use the event-stream filehandle from a child process will
throw an exception.

We have also suffered a lot from double-encoding utf8, specially in
JSON. I have read up on how the JSON flags work and I have added
->ascii(1) to the JSON encoders. This does not stop them from being
UTF8, it just forces them to \u#### escape any codepoints above 127.
This avoids any double encoding in the JSON strings.