Use setpgrp in daemon mode, not in foreground
The logic on this parameter appears to have been inverted -- setpgrp
should be *used*, not omitted, when yath is running as a daemon.

The misbehavior that this caused was subtle. Most use cases worked fine;
the only exception was situations where yath was started from a terminal
which was closed immediately afterwards, e.g.

    ssh -t host yath start

Since yath didn't call setpgrp, it would remain associated with the
terminal, and would receive SIGHUP when the terminal was closed.