Fix coverage race condition
While all events for a job are ordered, events across multiple jobs can
be out of order. As such events from a retry can be seen before the
first run of a test file is fully processed. Retries can produce
duplicate coverage data, so in this race condition we can have a
duplicate entry. This change fixes that so only the final job/retry is
used for coverage data.