Make reload smarter
Sometimes, like with DBIx::Class, it is common to load the actual db
schema in one class, and use overlays to bring in other methods. If the
overlay has the regular package name, and the schema is loaded via an
auto-compiler or similar then reloading a package would delete all the
generated subs and only reload the overlay. This commit fixes that by
only removing subs defined in the file being reloaded.