Rework shared slots
The previous implementation did not play nicely with other shared
resources. Slots might get assigned to tests that could not run due to
other resourced being unavailable, this was very wasteful.

This implementation cares less about which tests get which slots. There
is still some tracking of that in order to know how many tests a run has
left, but specific tetss are not assigned slots until they are ready to
run.