Fix permissions issue with state files
The state files were being written using atomic-swap, which writes the
file to a temporary one (ext .pend) and then uses rename to override the
correct file with the temporary one. This caused issues when one person
owned the file and another tried to update it.

The fix was to switch it to not use atomic ones, it opens the original
with '>' and replaces the contents