Fix #293
Don't unlink during END for your tempfile when a tempdir is already
set to cleanup in END and passed in as the workdir.

Possibly I could be more intelligent about this and maybe check if
workdir is *not* matching `^yath-` (set by App::Yath::Tester), but
thus far I don't yet see a context where this wouldn't be true, so
it just doesn' seem worth checking for.