Fix #291 - NULL characters cannot be stored in postgresql
PostgreSQL does not allow null characters in json fields.

This fixes things for PostgreSQL by adding an extra escape to all null
characters just before decoding the json when using the run processor
and a PostgreSQL database. This does not apply to other databases, but
could if any end up with similar behavior.

I chose to escape them instead of stripping them as that makes it more
obvious when someone looks at it. I actually suspect we may want to
ALWAYS do this so that the null characters are always visible... But not
yet.

This will find any \0 or \u0000 that has no escapes (IE the string
literal is '\\0' or '\\u0000') And adds an escape. If there is already
another escape in front of it then it is not modified, this avoids
double-escaping any already escaped, and also will avoid adding escapes
to nested json structures.